render: debounce render requests to 4ms
The requestRender method sends to a channel and requests the host app to
redraw the terminal and call screen.Show. The requests are very
frequent, and are easily dropped if not using a large buffered channel.
Minimal testing showed that reaching 1000 in-queue requests was not
uncommon.

Debounce the requests to a reasonable refresh rate. 4 milliseconds is
250hz which should be plenty for TUIs.

Signed-off-by: Tim Culverhouse <tim@timculverhouse.com>