terminal: BREAKING: use a redraw poll instead of channel
The processing of runes creates a very large amount of data sent to the
redrawRequest channel. Use a bool and periodic check for if a render is
required.

Previously, a debounce had been added. The debounce functioned similarly
to this implementation however there are fewer chances for error in this
implementation.

Signed-off-by: Tim Culverhouse <tim@timculverhouse.com>