redraw: set redraw to false earlier
Set redraw to false prior to setting content to the view. This will
result in more calls to Draw, but we only draw dirty cells anyways.

We could use a mutex, however I think blocking access to the Terminal or
Buffer object will result in worse performance than forcing more calls
to Draw.

Setting redraw to false earlier means that a cell could be set dirty
during the draw loop, it gets drawn, but term.redraw is also set back to
true. This means we'll call Draw again the next tick, but that cell
still won't be redrawn because it's own methods will have set it to
dirty.

Signed-off-by: Tim Culverhouse <tim@timculverhouse.com>