terminal: fix up length of UTF-8 copy buffer
The current code tries to read into `buf[len(buf):cap(buf)]`. but at the
first `for-loop` iteration, `len(buf)==cap(buf)`, so
`len(buf[len(buf):cap(buf)]) == 0`. so we try to read into a 0-size
buffer, which isn't so useful.

Create the buffer with the correct capacity but with an initial
0-length. this should save an iteration.

Signed-off-by: Sebastien Binet <binet@cern.ch>
Reviewed-by: Tim Culverhouse <tim@timculverhouse.com>