grid: implement smaller interface for drawing
Define a smaller interface for the drawing surface, called a Grid. A
Grid only defines the methods used by tcell-term, and still allows for a
views.View and a tcell.Screen to be a Grid. By defining a smaller
interface, it can be easier for implementers to create their own custom
Grid to draw on.

Signed-off-by: Tim Culverhouse <tim@timculverhouse.com>