csi: limit Cursor Horizontal Absolute to positive ints
Cursor Horizontal Absolute escape codes set the absolute position of the
cursor on the current line. The sequence is '\x1b[<p>G', where <p> is
the column the cursor should be placed. If <p> is 0, the cursor is sent
to the first column.

Commit 690223c496ae ("types: clean up uintX type insanity") changed many
of the types to ints, which removed some implicit clamping to 0. This
introduced a regression where a <p> of 0 would request the position be
set to -1, resulting in an Out of Bounds panic.

Limit the position the cursor can be set in the setPosition method to
only positive integers.

Fixes: 690223c496ae ("types: clean up uintX type insanity")
Fixes: https://todo.sr.ht/~rockorager/tcell-term/5
Fixes: https://todo.sr.ht/~rjarry/aerc/133
Signed-off-by: Tim Culverhouse <tim@timculverhouse.com>